lib/amdgpu: Add user mode queue support in ring context
Add the meta data to support the user mode command
submission in the ring_context.

User mode command submission methods needs these
resources to be initialized and to create/destroy queues.

Also once we have the queue created the queue id is
used to submit the work load to the h/w.

Signed-off-by: Sunil Khatri <sunil.khatri@amd.com>
Reviewed-by: Vitaly Prosyak <vitaly.prosyak@amd.com>