runner/executor: Use bootime for time checks
Runner is printing time to stdout/stderr with CLOCK_BOOTTIME but
when measuring time for a test it is using an igt library which
uses CLOCK_MONOTONIC. This leads to an errors when calculating
time left for tests using suspend or hibernate. For example log:

[77.376851] [020/123] (932s left) kms_flip (2x-flip-vs-suspend)
[77.592412] Starting subtest: 2x-flip-vs-suspend
[77.604996] Starting dynamic subtest: AB-DP2-HDMI-A3
[115.135795] Dynamic subtest AB-DP2-HDMI-A3: SUCCESS (6.776s)

shows test runs for around 6 seconds, while wall time shows
it took around 38 seconds.

  Create a separate, runner specific function for time measure
and use it where current time is needed.

Cc: Ewelina Musial <ewelina.musial@intel.com>
Cc: Petri Latvala <adrinael@adrinael.net>
Reported-by: Karol Krol <karol.krol@intel.com>
Signed-off-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Reviewed-by: Dominik Karol Piątkowski <dominik.karol.piatkowski@intel.com>
Tested-by: Peter Senna Tschudin <peter.senna@linux.intel.com>
Reviewed-by: Peter Senna Tschudin <peter.senna@linux.intel.com>