tests/kms_universal_plane: Fix fb count mismatch
There is a mismatch in the count of fb before
start and end of the test. This is because cursor
fbs still not removed and we are counting the fbs.
Add a 2 vblank wait before counting the fbs so that
all cursors fbs gets removed.
Since unpinning of cursor fb occurs during vblank in xe,
we need to wait atleast 1 vblank for driver to remove
cursor fb. We need 1 additional vblank because vblank
event is scheduled before unpinning.

v2: Use is_xe_device[Chaitanya]
v3: Add closes tag[Swati]

Closes: https://gitlab.freedesktop.org/drm/xe/kernel/-/issues/4415
Signed-off-by: Nemesa Garg <nemesa.garg@intel.com>
Reviewed-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>