test/intel/prime_mmap: Limit aperture testing to machines with an aperture
The test_aperture_limit looks at the handling of internal GGTT mmaps for
the dmabuf and checking what happens if we try to mmap more than we can
fit inside the aperture. Thus if we have no aperture, there is no corner
case regarding the mmapable aperture that requires validation.

Signed-off-by: Chris Wilson <chris.p.wilson@linux.intel.com>
Signed-off-by: Pravalika Gurram <pravalika.gurram@intel.com>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>