tests/intel/xe_fault_injection: Return result from inject_fault_probe()
In the current implementation, test asserts if the captured error is not same as
injected error. It is possible that the error received is translated to other
error which can be returned to application and in some scenarios driver retries
in case of failure and so, no error might be captured.

Considering above cases, not asserting the test in inject_fault_probe(), but
returning error to the caller. Caller can check and assert if needed.

Signed-off-by: Satyanarayana K V P <satyanarayana.k.v.p@intel.com>
Reviewed-by: Marcin Bernatowicz <marcin.bernatowicz@linux.intel.com>