lib/amdgpu: fix potential issue with bitwise AND operation
The availability of reset mask options should be tested using a
bitwise AND, but since we’ve only ever used a single option per
call, this typo didn’t manifest as a bug.
Also, improve the debug message to provide more clarity when
the requested reset type is not supported by the driver or firmware.

Cc: Jesse Zhang <jesse.zhang@amd.com>
Signed-off-by: Vitaly Prosyak <vitaly.prosyak@amd.com>
Reviewed-by: Jesse Zhang <jesse.zhang@amd.com>