tests/intel/xe_exec_system_allocator: Check huge pages free in thp tests
In a previous patch, various subtests in xe_exec_system_allocator that
rely on the MAP_HUGETLB flag for mmap were skipped in the event that no
huge pages were reserved by the kernel.  These tests should also be
skipped if all the huge pages are currently in use, which is indicated
by HugePages_Free being 0 in /proc/meminfo.

Fixes: df3036da ("tests/intel/xe_exec_system_allocator: Check for huge pages before thp tests")
Signed-off-by: Jonathan Cavitt <jonathan.cavitt@intel.com>
Cc: Matthew Brost <matthew.brost@intel.com>
Cc: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Reviewed-by: Matthew Brost <matthew.brost@intel.com>