tests/intel/xe_mmap: Acquire rpm reference for pci-membarrier tests
During the test, the device is entering into runtime suspend, which is
leading the device to d3hot or d3cold mode. Hence the read is giving
a special value 0xffffffff.

To avoid this issue acquired an rpm ref to keep the device awake
during the access.

Also created a sub-group for all the pci-membarrier tests.

Closes: https://gitlab.freedesktop.org/drm/xe/kernel/-/issues/4784
Signed-off-by: Kolanupaka Naveena <kolanupaka.naveena@intel.com>
Reviewed-by: Matthew Auld <matthew.auld@intel.com>