tests/intel/xe_configfs: Test engines_allowed
Add tests for the parsing of engines_allowed configfs. This doesn't try
to bind the driver and ensure only those engines are available. That's
because the engine uapi instance doesn't necessarily match what is used
for configfs configuration. Maybe we will need an engine map in debugfs
for covering that.

Reviewed-by: Umesh Nerlige Ramappa <umesh.nerlige.ramappa@intel.com>
Reviewed-by: Riana Tauro <riana.tauro@intel.com>
Link: https://lore.kernel.org/r/20250602-engines-allowed-v2-2-52bffe717766@intel.com
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>