include/android: Add stub for glib
As glib is not available in Android, not able to build igt for Android.

Provided stub implementations so igt can be compiled for both Android
and linux from same code base. In case of linux systems, system
libraries will be used whereas for Android stubs will be used with
reduced test coverage.

Signed-off-by: Sapna Singh <sapna1.singh@intel.com>
Signed-off-by: Jeevaka Prabu Badrappan <jeevaka.badrappan@intel.com>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>