tests/intel/xe_exec_system_allocator: Add prefetch sections
Update the test_exec() function with flags to initiate and understand
prefetches. Prefetching is tested by verifying that the fault count in
the GT stats has not  increased. However, a mismatch does not cause the
test to fail—instead, a warning  is printed. This is due to the
unreliability of system allocator behavior; the  added sections
primarily serve to ensure that the driver remains stable.

That said, if the page fault counts are consistently incorrect, it may
indicate a  problem in the KMD.

A prefetch benchmark section has also been added to help profile
prefetch behavior  in the KMD.

v2:
 - Prefetch initial buffer on SRAM devices (CI)

Signed-off-by: Matthew Brost <matthew.brost@intel.com>
Reviewed-by: Himal Prasad Ghimiray <himal.prasad.ghimiray@intel.com>