tests/kms_flip: Track and report unexpected events with pass rate metric
At higher refresh rates, there is a possibility of flips being
submitted during the evasion window, which can result in one extra
frame and cause the test to fail. To address this, the test has
been modified to calculate a pass rate based on the number of
expected versus erroneous frames. A threshold of 85% is enforced
to determine test success.

v1: Patch #2 and #4 merge into single patch. (Karthik)
v2: Address the regression failures.

Signed-off-by: Naladala Ramanaidu <ramanaidu.naladala@intel.com>
Reviewed-by: Karthik B S <karthik.b.s@intel.com>
Reviewed-by: Arun R Murthy <arun.r.murthy@intel.com>