runner/settings: Fix while loop
Passing a large argument to the --hook option caused igt_runner to
crash.  Valgrind identified an issue in read_hook_strs_from_file(),
which this patch addresses.

A typo in parenthesis placement caused the assignment of 0 or 1 to
line_length, not the actual line length, because != -1 was evaluated
before assignment.

Cc: Gustavo Sousa <gustavo.sousa@intel.com>
Signed-off-by: Peter Senna Tschudin <peter.senna@linux.intel.com>
Reviewed-by: Gustavo Sousa <gustavo.sousa@intel.com>