tests/msm/msm_mapping: Add wait for stall-on-fault to re-arm
Newer msm driver disable stall-on-fault for a grace period, to avoid
a flood of faults causing instability with memory translations that
the hw attempts with the translation stalled.  Fortunately it adds a
debugfs file so we can know how long we need to wait for stall-on-
fault to be re-enabled.

See: https://lore.kernel.org/all/20250520-msm-gpu-fault-fixes-next-v7-0-96cd1cc9ae05@gmail.com/

Signed-off-by: Rob Clark <rob.clark@oss.qualcomm.com>
Acked-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>