lib/xe: Don't assume GT ID matches location in GT query list
To prevent confusion between GT ID (uapi) and what a GT's index in the
list returned by the query ioctl, switch some internal functions to take
a pointer to the drm_xe_gt structure itself.  Also add a helper function
that can find this structure in the GT query list, given a uapi GT ID.
Current platforms always have list index == uapi ID, but this is not
guaranteed to be true for future platforms.

v2:
 - Switch some ffs() to ffsll() while making these changes to ensure the
   64-bit masks are handled properly on 32-bit architectures. (Zbigniew)

Cc: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Reviewed-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>