runner: Relax timeout reduction on soft lockup
In case of soft lockups, it might be helpful from root cause analysis
perspective to see if the test was still able to complete despite
triggering the soft lockup warning, or if that soft lockup seems not
recoverable without killing the test. For that to be possible, igt_runner
should not kill the test too promptly if a soft lockup related kernel
taint is detected.

On kernel taints, igt_runner now decreases per test and inactivity
timeouts by a factor of 10.  Let it check if the taint is caused by a
soft lockup and decrease the timeouts only by the factor of 2 in those
cases.

v2: Define symbols for taint bits and use them (Kamil)

Signed-off-by: Janusz Krzysztofik <janusz.krzysztofik@linux.intel.com>