tests/intel: Remove intel_sysfs_debugfs
The intel_sysfs_debugfs test was originally introduced to unify testing
for both i915 and Xe drivers, aiming to prevent gaps in test coverage.
However, the implementation fell short in addressing scenarios where
both i915 and Xe were present in the system.

To address these limitations, it was replaced by:
 - tests/core_debugfs (gpu agnostic)
 - tests/core_sysfs (gpu agnostic)
 - tests/intel/xe_debugfs (complementary to core_debugfs)

Cc: kamil.konieczny@linux.intel.com
Cc: katarzyna.piecielska@intel.com
Cc: zbigniew.kempczynski@intel.com
Cc: michal.wajdeczko@intel.com
Cc: karthik.b.s@intel.com
Cc: ewelina.musial@intel.com
Reviewed-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Signed-off-by: Peter Senna Tschudin <peter.senna@linux.intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>