lib/amdgpu: reduce the job count for deadlock helpers
The current limit (200), takes significantly longer with
the improved reset code which the new code just kills the
bad job, not all jobs in the queue.  This leads to the
deadlock tests taking longer than expected and getting
killed by the CI systems for taking too long.  Reduce
the limit to 20.

Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
Reviewed-by: Vitaly Prosyak <vitaly.prosyak@amd.com>
Change-Id: I93d27c00c328a2df8762ffefcfc5c483d5473903