lib/intel/veboxcopy: Respect buffer MOCS index
Set the MOCS indices in surface state from the source and
destination buffers. So far it was done only for the input
buffer and only for 4Tile formats. Always setting it alings
with how rendercopy does it and it fixes the AuxCCS tests on
ADL when target is of the MC flavour.

Signed-off-by: Tvrtko Ursulin <tvrtko.ursulin@igalia.com>
Cc: Imre Deak <imre.deak@intel.com>
Cc: Juha-Pekka Heikkilä <juha-pekka.heikkila@intel.com>
Cc: Mika Kahola <mika.kahola@intel.com>
Reviewed-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>