tests/intel/kms_psr: Do not test frontbuffer rendering on cursor plane
Cursor doesn't generally support frontbuffer rendering. This test happens
to work on i915 because i915 has frontbuffer flush call on render
completion. We do not have that in Xe.

On Xe this mostly happens to work currently because any plane/pipe register
write is triggering update and that is keeping display and PSR awake. Now
as we are moving to using trans push mechanism to trigger update this
doesn't work anymore.

Stop testing cursor frontbuffer rendering in kms_psr if using Xe driver.

Signed-off-by: Jouni Högander <jouni.hogander@intel.com>
Reviewed-by: Swati Sharma <swati2.sharma@intel.com>