lib/xe_gt: Assertion related to xe_gt_stats_get_count
Currently, when the specific gt stats file is not found, it fails with
an assertion. Due to this the test using this function also fails.
In general, this lib function is just to count the stats count, test
using this need not fail.So, now the function will warn the failure
when file not found and return -1 as count.

Signed-off-by: Priyanka Dandamudi <priyanka.dandamudi@intel.com>
Reviewed-by: Nishit Sharma <nishit.sharma@intel.com>