lib/igt_pm: Skip LPM policy control on unsupported SATA hosts
With [1], LPM policy control is now disabled for external SATA ports.
This led to a regression in our CI runs [2]. Discussions on the regression
resulted in patch [3] which introduces a new sysfs entry, querying
which we can determine if LPM policy control is allowed on a host or not.

Store/Re-store LPM policy only when it is allowed. If the sysfs entry is
not present assume LPM policy control is allowed. This  maintains backward
compatibility.

[1] https://lore.kernel.org/linux-ide/20250701125321.69496-8-dlemoal@kernel.org/
[2] https://lore.kernel.org/linux-ide/07563042-6576-41cd-9a95-de83cfc95de1@intel.com/
[3] https://lore.kernel.org/linux-ide/20250730001947.332661-1-dlemoal@kernel.org/

Signed-off-by: Chaitanya Kumar Borah <chaitanya.kumar.borah@intel.com>
Reviewed-by: Damien Le Moal <dlemoal@kernel.org>