tests/intel/xe_fault_injection: Break early if the test is done
For the GuC communication injections, the test is currently looping for
a hardcoded 100 times. In the current driver we call the 2 GuC
communication functions way less than 100 times each (in one run I've
counted less than 40 for one and less than 10 for the other), so for
the majority of the iterations the test is just reloading the driver
without any injection occurring.

To avoid wasting so much time, we can just break early if we've reached
the point where no more injections are occurring. If we use a positive
value for the "times" parameter in the injection framework, the value
will be decremented every time an injection occurs in the given
iteration, so we can break once we detect that the value is unchanged.

While at it, add an assert to alert us if the number of calls to the
function increases above our current maximum number of loops, so we know
that the test needs to be updated.

Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Cc: Satyanarayana K V P <satyanarayana.k.v.p@intel.com>
Reviewed-by: Jonathan Cavitt <jonathan.cavitt@intel.com>