tests/intel/xe_fault_injection: Add test for xe_pxp_exec_queue_add
This test will make sure that the error path in the exec_queue creation
code works correctly if there is a PXP error.

Since this is the first queue creation case that requires an extension,
the code has been updated to add support for that

This patch requires the matching kernel code [1] to be merged first.

[1] https://patchwork.freedesktop.org/series/153095/
Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Reviewed-by: John Harrison <John.C.Harrison@Intel.com>