tests/intel/xe_configfs: Do not execute with open fd
The survivability_mode test does an unbind/set-configfs/bind sequence.
If and fd is open, that will still refer to the old instance of the
driver and is a dangling fd. Close it right when it's not needed anymore
as in future other tests will also unbind/bind.

While moving this code around, also make sure to use drm_close_driver()
so the fd is dropped from the cache.

Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Link: https://lore.kernel.org/r/20250825-configfs-unbind-v2-1-62e7508d3ed7@intel.com
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>