intel-ci/xe-sriov-*: Extend VF and VM blocklists
Currently VF and VM blocklists are supposed to be short and only
contain tests that are breaking/aborting the CI run. However
with the huge amount of expected skips present in VM runs it is
justified to change the approach of using those blocklists and
include all tests that are expected to skip when driver is in VF
mode.
Extend VF and VM blocklists and include expected skips. This
should speed up CI runs in configurations where driver is in VF
mode (mainly VM runs where many kms tests are skipping).

Signed-off-by: Jakub Kolakowski <jakub1.kolakowski@intel.com>
Cc: Adam Miszczak <adam.miszczak@linux.intel.com>
Cc: Marcin Bernatowicz <marcin.bernatowicz@linux.intel.com>
Cc: Lukasz Laguna <lukasz.laguna@intel.com>
Cc: Satyanarayana K V P <satyanarayana.k.v.p@intel.com>
Reviewed-by: Marcin Bernatowicz <marcin.bernatowicz@linux.intel.com>