tests/kms_hdr: Set HDR meta-data while testing for panel capability
Setting max_bpc simply configures the pipeline to drive
10 bits per color channel, but HDR requires more than higher bit
depth. HDR also depends on metadata such as EOTF, peak luminance,
and colorimetry.

Update the test to set HDR metadata on the connector to signal that
we need an HDR pipeline, before running through the modes. Reset
it afterwards.

If setting HDR metadata fails for all advertised modes on a supported
panel, it implies the kernel cannot drive HDR — for example, due to
bandwidth limitations.

v2:
	- Update commit message (Swati)
	- Add comment on why we set HDR meta data (Swati)
	- Commit after setting HDR meta data (Swati)

v3:
	- Handle error while committing meta data (Swati)

Signed-off-by: Chaitanya Kumar Borah <chaitanya.kumar.borah@intel.com>
Reviewed-by: Swati Sharma <swati2.sharma@intel.com>