tests/kms_async_flips: Optimise subtests for simulation platforms
Restrict the crc,dpms,hang subtests to execute on a single pipe on
the simulation platforms as the crc calculations are slow.
This avoids timeouts on CI during execution while ensuring
functionality of the test is verified with single pipe.

V2: Update commit message, refactor logic to use single pipe (Karthik)

Signed-off-by: Santhosh Reddy Guddati <santhosh.reddy.guddati@intel.com>
Reviewed-by: Karthik B S <karthik.b.s@intel.com>