tests/intel/kms_ccs: stop using l3 enabled PAT index
When populating the fb using an l3 compression enabled PAT index some or
all of the data may end up cached. However the HW looks to only perform
the compression step in this case upon evicting those entries, when also
trying to write them back to VRAM. This seems to be the cause of this
test randomly failing on BMG with the CCS state appearing to sometimes
be zeroed even after doing a compression enabled copy. From
experimentation adding a sleep before the surface copy cures the
failure, which fits since this will give plenty of time to enter rc6
which will indirectly also nuke the l3.  Grabbing a forcewake around the
sleep brings back the failure which also makes sense since this will
inhibit the flush and also rules out missing synchronisation hidden by
the sleep. Using a large fb also cures the issue, which also fits since
the fb will now be larger than the entire l3, so some data will have to
be compressed when evicted.

To fix this don't use an l3 enabled PAT index prior to taking a snapshot
of the raw CCS state. Probably this also means the test is maybe too
much looking at implementation details, by assuming that zeroed CCS
state must also imply that there is no compression, even if compression
is merely delayed until the data is evicted.

Closes: https://gitlab.freedesktop.org/drm/xe/kernel/-/issues/5941
Closes: https://gitlab.freedesktop.org/drm/xe/kernel/-/issues/5376
Signed-off-by: Matthew Auld <matthew.auld@intel.com>
Cc: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Cc: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>
Reviewed-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>