tests/intel/xe_configfs: Check ctx_restore_post_bb
Add sanity checks for some bb written to ctx_restore_post_bb and
ctx_restore_mid_bb. For invalid values, it validates the xe configfs
implementation correctly rejects them. For valid values, it checks the
values are written correctly and that later when we driver binds that
the HW register contains the expected value.

Reviewed-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Link: https://lore.kernel.org/r/20250918-configfs-ctx-restore-v2-2-5df035e27e42@intel.com
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>