tools/intel_display_poller: Add --scanline-offset/-o command line option
Our hardware has a slightly warped view of scanline numbers.
Eg. on most platforms it reports DSL=vblank_start-1 when the
start of vblank actaully occurs.

On the kernel side we have helpers to convert between the hardware
and software scanline numbers. Add the same for intel_display_poller
in the form of a new --scanline-offset/-o command line option. With
this one doesn't constantly have to waste brarin power to convert the
hardware numbers into something more sensible.

Caveat: unlike the kernel side we don't do the appropriate %vtotal
on the adjusted values (since figuring out the actual vtotal would take
quite a bit of work due to VRR/etc.). So when using --scanline-offset
one has to keep in mind that we might see slightly funky scanline
numbers. Eg. scanline >= vtotal when the scanline offset of positive,
and negative value (due to a negative scanline offset) we just
saturate to 0. Howver for most interesting hardware events the
scanline tends to be in a range where this issue doesn't matter.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Reviewed-by: Michał Grzelak <michal.grzelak@intel.com>