tests/kms_plane: Add test to validate odd panning
This feature creates a NV12 plane having two framebuffers
and moving it to odd position and then do continuous flip
along with the change in the width of framebuffer.
This is to check whether panning at odd position is
supported or not.

v2: Add test in kms_plane.[Swati]
v3: Remove extra parameters[Pranay]
v4: Add check for modifier
v5: Adding the break condition
v6: Change the fb size
v7: Make it a part of planar_seeting[JP]

Signed-off-by: Nemesa Garg <nemesa.garg@intel.com>
Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>