tools/intel_vbt_decode: Introduce to_str()
Introduce a to_str() macro that gives us a standard way
to decode a value we got from the VBT into a string.

All the cases converted here were implemented as a custom
function to deal with out of bounds/sparse values. All of
that is now taken care of by to_str(), so the strings can
just be defined as an array.

v2: Also catch negative indices (Kamil)

Reviewed-by: Jani Nikula <jani.nikula@intel.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>