lib/intel/intel_compute: Add kernel validation for compute IP versions
Currently, there are no comprehensive checks for the availability of all
the required kernel types in compute kernel. This patch introduces
helper functions to perform kernel lookup and validation based on the
provided IP version.

The validation function ensures that the necessary kernels (square,
SIP, loop, long) are available depending on the preemption type.

Signed-off-by: Sobin Thomas <sobin.thomas@intel.com>
Reviewed-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Link: https://lore.kernel.org/r/20251013133439.543393-3-sobin.thomas@intel.com