tests/intel/gem_eio: Skip on too few reset-resume cycles
Subtests that measure time of resume after engine reset compare a median
value calculated from the measurements against a presumed limit and fail
if the limit has been exceeded.  However, if it occurs not possible to
collect enough measurements required for stable median value calculation,
that condition is now ignored and success is reported, as if the measured
time fit below the limit.

Skip if not able to collect sufficient number of time measurements.  CI
results from slow platforms that always skip may be handled as expected
skips.

v3: Use a symbol for minimum number of measurements (Kamil),
  - reword skip message for its better readability (Kamil).

Cc: Krzysztof Karas <krzysztof.karas@intel.com>
Cc: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Signed-off-by: Janusz Krzysztofik <janusz.krzysztofik@linux.intel.com>
Reviewed-by: Krzysztof Karas <krzysztof.karas@intel.com>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>